Add files via upload
GDELT daily parser takes data from every day in a single year, splits the events of that day into 24 hrs, and adds all the hours for the entire year into a single file. The monthly parser does something similar but with the earlier monthly format

SpreadDataOver3Days does what the title says. In the case of missing days, the data for that day is taken from the neighboring days